Move the fingerprint tests into fetch.tls.test.ts
Also buffer the CONNECT request to its blank line in the tunnel test,
document the spare-capacity contract of zstd decompress_append and drop
the two TLS version constants nothing reads anymore.
